bug fix reboot

+++ src/00_install.sh
@@ -10,16 +10,21 @@ if [[ $M_INSTALL_ANSWER =~ ^[Yy]$ ]]; then
 
     source 03_base_install.sh
 
-    source 034_grub_install.sh
+    source 04_grub_install.sh
 
-    source 04_config_install.sh
+    source 05_config_install.sh
 
-    source 05_recommanded_package_install.sh
+    source 06_recommanded_package_install.sh
 
-    source 06_graphical_interface_install.sh
+    source 07_graphical_interface_install.sh
 
-    source 07_create_user_install.sh
+    source 08_create_user_install.sh
 
     echo "Remove the installation media."
-    reboot
+
+    read -p "Installation finished, do you want to reboot ? [Y/n]" M_REBOOT
+
+    if [[ $M_REBOOT =~ ^[Yy]$ ]]
+        reboot
+    fi
 fi
